Move Logger to logger/Logger to avoid same-name-modulo-upper-case issue
