+++ lib/logger.es6.js
@@ -23,7 +23,7 @@ const fs = require('fs');
 const BoundLogger = require('./logger/BoundLogger.es6.js');
 const ColorConsoleLogger = require('./logger/ColorConsoleLogger.es6.js');
 const FileLogger = require('./logger/FileLogger.es6.js');
-const Logger = require('./Logger.es6.js');
+const Logger = require('./logger/Logger.es6.js');
 const PrefixLogger = require('./logger/PrefixLogger.es6.js');
 const SplitLogger = require('./logger/SplitLogger.es6.js');
