+++ lib/logger/FunctionLogger.es6.js
@@ -16,7 +16,7 @@
  */
 'use strict';
 
-const Logger = require('../Logger.es6.js');
+const Logger = require('./Logger.es6.js');
 
 class FunctionLogger extends Logger {
   data() {
