+++ lib/logger/Logger.es6.js
@@ -16,7 +16,7 @@
  */
 'use strict';
 
-const NullLogger = require('./logger/NullLogger.es6.js');
+const NullLogger = require('./NullLogger.es6.js');
 
 class Logger {
   // Re-implement part of Base, but Base requires this (so cannot inherit from
